Implementado Endpoint Subir Pedido(post /pedidos) Consumo da Api Pix Efipay.

+++ src/main/java/com/app/ecommerce/dto/CartItemDto.java
@@ -1,5 +1,5 @@
 package com.app.ecommerce.dto;
 
-public record CartItemDto(Integer id_prod, Integer quantity) {
+public record CartItemDto(Long id_prod, Integer quantity) {
 
 }
